Thermal Meter: use strchr to find the terminating newline of the zone type
Apparently some systems read two newlines from the 'type' sysfs
attribute, but I was only checking the last character of the buffer,
leading to a '?' after the name.

Terminate the string at the first newline to fix this.

Also make sure the buffer is large enough for the trailing nullbyte and
check the return value of fread to avoid warnings